Schedule Data for AY 2024-2025

This dataset contains the schedule data for the academic year 2024-2025.

Usage

schedule_ay24_25

Format

A data frame with 20 columns:

objectID

Unique identifier for each course

year

Academic year

subject

Subject of the course

code

Course code

title

Course title

description

Course description

units_min

Minimum number of units

units_max

Maximum number of units

term

Term

term_id

Term ID

section_number

Section number

component

Component

class_id

Class ID

current_size

Current class size

max_size

Maximum class size

days

Days of the week

start_time

Start time

end_time

End time

location

Location

department

Department

Source

Data is from the Office of the University Registrar